ASR reserching log
This researching log will describe the whole process of the development of the ASR project, especially the embedded electronic control system which I lead and participate from the beginning of the project. In the other words, this log is the record of things I have beset with during the research. So, this log is not strictly in the form of the formal paper and it should be just considered as the simple memorandum which is used as the raw material for the following paper writing.
This log will follow the order of researching time. Each chapter will state one progress or the mistake that I have made or faced. In the last chapter I will write about the final model of all the components of the whole project.